Remove duplicate class-based components
